Move a monitor_indexes::Db::latest_schema_version into a db actor
This is a part of #3.

The change moves latest_schema_version method from monitor_indexes into a db
actor. Next paches will move other methods into a db actor.
